Common.cpp: fix LastDelimiter

+++ src/core/Common.h
@@ -241,7 +241,7 @@ size_t AnsiPos(const std::wstring &str, wchar_t c);
 size_t Pos(const std::wstring &str, const std::wstring &substr);
 std::wstring StringReplace(const std::wstring &str, const std::wstring &from, const std::wstring &to);
 bool IsDelimiter(const std::wstring &str, const std::wstring &delim, size_t index);
-int LastDelimiter(const std::wstring &str, const std::wstring &delim);
+size_t LastDelimiter(const std::wstring &str, const std::wstring &delim);
 //---------------------------------------------------------------------------
 
 int CompareText(const std::wstring &str1, const std::wstring &str2);
